feat: move useEditableText to hooks
